Tidy up UX
Moved most existing logging over to `tracing` on `debug` or `trace`
levels.

While doing this I found we had no format layer set up on `tracing` so I
added that.

Also tidied up the output when it reads from `cargo` and we now output
that formatted with `owo-colors` to stdout.