Union `package.metadata.fsm.build-inputs` fields
In an effort to embed native dependencies into a `Cargo.toml`, we read
from the `package.metadata.fsm.build-inputs` table for a list of package
names (as keys) with semver markers for required version.

Currently versions are ignored, this is a TODO.

Package dependencies are unioned together to create a full build-inputs
list.