Switch the loading unicode symbols to, uh, different ones
The old ones would smush on top of each other when rendered by
fancier renderers. These seem to render better, but it may be
that we just can't use unicode in our loader.

Another change is I made the tick interval 260ms to match the
high-hat I have in my head:

(115 beats per minute) / (60 seconds) = .52s per quarter note
= .26 seconds per eighth note
= 260ms per eighth note

Note that I also double-render the quarter note so it is
"played" for the correct amount of time.